float64 to 32

+++ src/tasks/memorymaze.py
@@ -1,4 +1,5 @@
 import gymnasium as gym
+import numpy as np
 import os
 
 # to enable rendering on a headless machine
@@ -11,13 +12,13 @@ class FrameProcessing(gym.Wrapper):
     def reset(self, **kwargs):
         obs, info = self.env.reset(**kwargs)
         self._last_frame = obs.copy()
-        obs = obs.astype(float) / 255
+        obs = obs.astype(np.float32) / 255
         return obs, info
     
     def step(self, action):
         obs, reward, done, truncated, info = self.env.step(action)
         self._last_frame = obs.copy()
-        obs = obs.astype(float) / 255
+        obs = obs.astype(np.float32) / 255
         return obs, reward, done, truncated, info
     
     def render(self):
